README
alias is the most powerful tool present in linux.
It allows you to create your own aliased commands in order to reduce your time.
The command created by you must not have the same name as the other files present in /usr/bin/ or /usr/sbin/.
If you are not satisfied with your command you have the power to overwrite the command or you can delete your command.
As soon as you create your command if you close the window the command gets removed. In order to make the command permanent
you must enter the command in the /.bashrc. 



<code>
//create new command where the command name is datum[i.e, date in german]
//the following command access both the date as well as cal command.

alias datum="date;cal"

//to remove the command from the terminal

unalias datum